+++ anknotes/html.py
@@ -1,6 +1,7 @@
 from HTMLParser import HTMLParser
-from anknotes.shared import get_evernote_account_ids
+from anknotes.constants import SETTINGS
 from anknotes.db import get_evernote_title_from_guid
+
 
 class MLStripper(HTMLParser):
     def __init__(self):
@@ -156,3 +157,43 @@ evernote_link_colors = {
   }
 }
 
+evernote_link_colors['Default'] = evernote_link_colors['Links']['Outline']
+evernote_link_colors['Links']['Default'] = evernote_link_colors['Default']
+
+enAccountIDs = None
+
+
+def get_evernote_account_ids():
+    global enAccountIDs
+    if not enAccountIDs:
+        enAccountIDs = EvernoteAccountIDs()
+    return enAccountIDs
+
+
+class EvernoteAccountIDs:
+    uid = '0'
+    shard = 's100'
+    valid = False
+    def __init__(self, uid=None,shard=None):
+        self.valid = False
+        if uid and shard:
+            if self.update(uid, shard): return
+        try:
+            self.uid =  mw.col.conf.get(SETTINGS.EVERNOTE_ACCOUNT_UID, SETTINGS.EVERNOTE_ACCOUNT_UID_DEFAULT_VALUE)
+            self.shard = mw.col.conf.get(SETTINGS.EVERNOTE_ACCOUNT_SHARD, SETTINGS.EVERNOTE_ACCOUNT_SHARD_DEFAULT_VALUE)
+        except:
+            self.uid = SETTINGS.EVERNOTE_ACCOUNT_UID_DEFAULT_VALUE
+            self.shard = SETTINGS.EVERNOTE_ACCOUNT_SHARD_DEFAULT_VALUE
+            return
+
+    def update(self, uid, shard):
+        if not uid or not shard: return False
+        if uid == '0' or shard == 's100': return False
+        try:
+            mw.col.conf[SETTINGS.EVERNOTE_ACCOUNT_UID] = uid
+            mw.col.conf[SETTINGS.EVERNOTE_ACCOUNT_SHARD] = shard
+        except:
+            return False
+        self.uid = uid
+        self.shard = shard
+        self.valid = True
