+++ anknotes/shared.py
@@ -4,9 +4,9 @@ try:    from pysqlite2 import dbapi2 as sqlite
 except ImportError: from sqlite3 import dbapi2 as sqlite
 
 ### Anknotes Imports
+from anknotes.constants import *
 from anknotes.db import *
 from anknotes.logging import *
-from anknotes.constants import*
 from anknotes.html import *
 
 ### Anki and Evernote Imports
@@ -20,6 +20,7 @@ try:
     from evernote.edam.error.ttypes import EDAMSystemException, EDAMErrorCode, EDAMUserException, EDAMNotFoundException
 except: pass
 
+# __all__ = ['ankDB']
 
 def get_friendly_interval_string(lastImport):
     if not lastImport: return ""
@@ -46,37 +47,6 @@ class UpdateExistingNotes:
     
 class EvernoteQueryLocationType:
     RelativeDay, RelativeWeek, RelativeMonth, RelativeYear, AbsoluteDate, AbsoluteDateTime = range(6)
-
-
-class EvernoteAccountIDs:
-    uid = '0'
-    shard = 's100'
-    valid = False 
-    def __init__(self, uid=None,shard=None):
-        self.valid = False 
-        if uid and shard:
-            if self.update(uid, shard): return 
-        try:
-            self.uid =  mw.col.conf.get(SETTINGS.EVERNOTE_ACCOUNT_UID, SETTINGS.EVERNOTE_ACCOUNT_UID_DEFAULT_VALUE)
-            self.shard = mw.col.conf.get(SETTINGS.EVERNOTE_ACCOUNT_SHARD, SETTINGS.EVERNOTE_ACCOUNT_SHARD_DEFAULT_VALUE)
-        except:
-            self.uid = SETTINGS.EVERNOTE_ACCOUNT_UID_DEFAULT_VALUE
-            self.shard = SETTINGS.EVERNOTE_ACCOUNT_SHARD_DEFAULT_VALUE
-            return 
-        
-    def update(self, uid, shard):        
-        if not uid or not shard: return False 
-        if uid == '0' or shard == 's100': return False 
-        try:
-            mw.col.conf[SETTINGS.EVERNOTE_ACCOUNT_UID] = uid
-            mw.col.conf[SETTINGS.EVERNOTE_ACCOUNT_SHARD] = shard
-        except:
-            return False 
-        self.uid = uid 
-        self.shard = shard 
-        self.valid = True 
-    
-enAccountIDs = None
 
 
 def get_tag_names_to_import(tagNames, evernoteTags=None, evernoteTagsToDelete=None):
@@ -123,19 +93,8 @@ def regex_see_also():
     if not _regex_see_also: update_regex()
     return _regex_see_also
     
-try:
-    icoEvernoteWeb = QIcon(ANKNOTES.ICON_EVERNOTE_WEB)
-    icoEvernoteArtcore = QIcon(ANKNOTES.ICON_EVERNOTE_ARTCORE)
-    imgEvernoteWeb = QPixmap(ANKNOTES.IMAGE_EVERNOTE_WEB, "PNG")
-    imgEvernoteWebMsgBox = imgEvernoteWeb.scaledToWidth(64)                
-except: pass
-
-evernote_link_colors['Default'] = evernote_link_colors['Links']['Outline']
-evernote_link_colors['Links']['Default'] = evernote_link_colors['Default']
 
 
-def get_evernote_account_ids():
-    global enAccountIDs
-    if not enAccountIDs:
-        enAccountIDs = EvernoteAccountIDs()
-    return enAccountIDs
+
+
+
